requirements.txt
torch~=2.2.2
torchvision~=0.17.2
pillow~=11.2.1
matplotlib~=3.10.1
future~=1.0.0
omegaconf~=2.3.0
transformers~=4.51.3
numpy~=1.26.4
tqdm~=4.67.1
opencv-python~=4.11.0.86
PyYAML~=6.0.2
scikit-image~=0.25.2
pytorch-lightning~=2.5.1
einops~=0.8.1
requests~=2.32.3
scipy~=1.15.2
setuptools~=78.1.0
diffusers~=0.33.1
gfpgan~=1.3.8